Make cmd global
I most likely don't edit chord files in the same nvim session as other
things so it's not really a problem to have a global cmd :Guitar when at
least one .crd file has been opened. If I wanted to be more correct
about things I would make it a module that gets required, then place the
call to make a buffer local cmd in the ftplugin file instead.