tex and typst buffer local keymaps
Checking that reverse search etc works for vimtex with sioyek.
It does for a small project like CV, but not for larger project. Maybe
issue with the project, maybe use skim in that case for latex. The
benefit of sioyek is no flicker mostly, but I think reverse search etc
is more important in a large project.