Only use nmap for finishing blockim
It takes no time to double press vs single press esc, so we don't really
need esc in insert mode to finish blockim. In some cases we won't be
done with the macro when exiting insert mode, but we will never normally
need esc in normal mode so it is a safe keybind.